backend/requirements.txt
fastapi>=0.95.0
pydantic>=2.0.0
uvicorn>=0.22.0
pyodbc==4.0.35
psycopg2-binary==2.9.6  # Avoid compilation issues
mysql-connector-python==8.0.33
torch==2.0.1
ollama